run/airflow/requirements.txt
google-cloud-automl>=0.1.2
